+++ src/main/java/chess/model/square/Square.java
@@ -1,6 +1,5 @@
 package chess.model.square;
 
-import chess.model.piece.Direction;
 import chess.model.piece.Color;
 import java.util.ArrayList;
 import java.util.List;
