+++ src/test/java/chess/model/square/DirectionTest.java
@@ -2,7 +2,6 @@ package chess.model.square;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.model.piece.Direction;
 import java.util.List;
 import org.junit.jupiter.api.Test;
